Implemented Phase 6 Sprint 7 todo item 6.7.8 - Restriction on the deletion of attachments.
	new file:   Configuration/DBScript/Phase6Sprint7.txt
	modified:   idas.chox.core/src/main/java/idas/chox/core/model/Attachment.java
	modified:   idas.chox.core/src/main/java/idas/chox/core/model/Claim.java
	modified:   idas.chox.core/src/main/java/idas/chox/core/services/AttachmentService.java
	modified:   idas.chox.data/src/main/java/idas/chox/data/services/AttachmentServiceImpl.java
	modified:   idas.chox.data/src/main/java/idas/chox/data/services/TaskServiceImpl.java
	modified:   idas.chox.data/src/main/resources/hbm/Attachment.hbm.xml
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/AssignWorkgroup.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/AttachmentAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/UserAction.java
	modified:   idas.chox.web/src/main/resources/applicationContext-services.xml

+++ idas.chox.data/src/main/java/idas/chox/data/services/TaskServiceImpl.java
@@ -98,7 +98,7 @@ public class TaskServiceImpl extends SecureDataService implements TaskService {
         //       false otherwise
         boolean canComplete = false;
 
-        if (task.getCreatedBy().getId() == webUser.getId()) {
+        if (task.getCreatedBy().getId().intValue() == webUser.getId().intValue()) {
             canComplete = true;
         } else if ((task.getVisibilityRole() == null || task.getVisibilityRole().length() == 0) && task.getVisibility()!= 1) {
             canComplete = true;
